Change of file to be build-specific
Sorry, not used to this kind of conf
