+++ lib/fs/basicfs_test.go
@@ -651,34 +651,6 @@ func TestXattr(t *testing.T) {
 	}
 }
 
-func TestRemoveWindowsDirIcon(t *testing.T) {
-
-	if !build.IsWindows {
-		t.Skip("only Windows")
-		return
-	}
-
-	fs, dir := setup(t)
-	relativePath := "folder_with_icon"
-	path := filepath.Join(dir, relativePath)
-
-	//Try to delete a folder with a custom icon with os.Remove (simulated by the readonly file attribute)
-	err := os.Mkdir(path, os.ModeDir)
-
-	a, e := syscall.UTF16PtrFromString(path)
-	if e == nil {
-		syscall.SetFileAttributes(a, uint32(syscall.FILE_ATTRIBUTE_DIRECTORY+syscall.FILE_ATTRIBUTE_READONLY))
-	} else {
-		t.Fatal(e)
-	}
-
-	err = fs.Remove(relativePath)
-
-	if err != nil {
-		t.Fatal(err)
-	}
-}
-
 func TestBasicWalkSkipSymlink(t *testing.T) {
 	_, dir := setup(t)
 	testWalkSkipSymlink(t, FilesystemTypeBasic, dir)
